% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAltSegments.R
\name{classifyAltSegments}
\alias{classifyAltSegments}
\title{Classify alternative splicing features}
\usage{
classifyAltSegments(transcript1, transcript2, txrevise_out = NULL)
}
\arguments{
\item{transcript1}{GRanges Object containing exon structure of transcript 1}

\item{transcript2}{GRanges Object containing exon structure of transcript 2}

\item{txrevise_out}{Optional.}
}
\value{
a list of GRanges objects similar to 'indentifyAddedRemovedRegions()' output (txrevise_out). 
?indentifyAddedRemovedRegions for more information
}
\description{
This function is an extension of the function indentifyAddedRemovedRegions by classifying
of the alternative segments
}
\examples{

classifyAltSegments(ptbp2Data$transcripts$ENSMUST00000029780, ptbp2Data$transcripts$ENSMUST00000197833)

}
