% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyAS.R
\name{classifyAS}
\alias{classifyAS}
\title{Classify alternative splicing between transcripts}
\usage{
classifyAS(tx1, tx2)
}
\arguments{
\item{tx1}{GRanges object of exons by transcript 1}

\item{tx2}{GRanges object of exons by transcript 2}
}
\value{
GRanges object of alternative segments
}
\description{
Classify alternative splicing between transcripts
}
