% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{.es_metric_symbol}
\alias{.es_metric_symbol}
\title{Symbolic notation for effect-size metric names, to keep columns narrow}
\usage{
.es_metric_symbol
}
\description{
Symbolic notation for effect-size metric names, to keep columns narrow
}
\keyword{internal}
