% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{.run_independent_anova}
\alias{.run_independent_anova}
\title{Execute an independent ANOVA-family test}
\usage{
.run_independent_anova(
  y,
  grp,
  normality_violated,
  variance_violated,
  test_type,
  warnings_list
)
}
\description{
Execute an independent ANOVA-family test
}
\keyword{internal}
