% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{.run_independent_two_sample}
\alias{.run_independent_two_sample}
\title{Execute an independent two-sample test}
\usage{
.run_independent_two_sample(
  y,
  groups,
  grp,
  normality_violated,
  variance_violated,
  test_type,
  warnings_list
)
}
\description{
Execute an independent two-sample test
}
\keyword{internal}
