% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{.run_paired_two_sample}
\alias{.run_paired_two_sample}
\title{Execute a paired two-sample test}
\usage{
.run_paired_two_sample(
  y,
  groups,
  grp,
  normality_violated,
  test_type,
  warnings_list
)
}
\description{
Execute a paired two-sample test
}
\keyword{internal}
