% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBestRef.R
\name{getBestRef}
\alias{getBestRef}
\title{Title}
\usage{
getBestRef(queryID, refID, gene_id, NMDer_ID, inputExonsbyTx,
  basicExonsbyTx, basicExonsbyCDS)
}
\arguments{
\item{queryID}{}

\item{refID}{}

\item{gene_id}{}

\item{NMDer_ID}{}

\item{inputExonsbyTx}{}

\item{basicExonsbyTx}{}

\item{basicExonsbyCDS}{}
}
\value{
df
}
\description{
Title
}
