% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{getORF}
\alias{getORF}
\title{_ workflow: Test transcript for NMD feature against a reference CDS}
\usage{
getORF(knownCDS, queryTx, refsequence, gene_id, transcript_id)
}
\arguments{
\item{knownCDS}{}

\item{queryTx}{}

\item{refsequence}{}

\item{PTC_dist}{}

\item{nonClassicalNMD}{}
}
\value{

}
\description{
THIS FUNCTION IS PART OF THE _ PROGRAM WORKFLOW.
This function will compare the query transcript with a reference CDS and tests
if insertion of alternative segments into the CDS generates an NMD substrate
}
