% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3togtfconvert.R
\name{gff3togtfconvert}
\alias{gff3togtfconvert}
\title{GFF3 to GTF converter}
\usage{
gff3togtfconvert(gffGRanges, comment = "")
}
\arguments{
\item{gffGRanges}{Input GFF3 GRanges object}
}
\value{
Outputs GTF GRanges object with mandatory gene_id and transcript_id headers
}
\description{
GFF3 to GTF converter
}
