% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunctions.R
\name{matchGeneIDs}
\alias{matchGeneIDs}
\title{Match Gene IDs from query GTF/GFF3 file}
\usage{
matchGeneIDs(query, ref, primary_gene_id = NULL, secondary_gene_id = NULL,
  outputfile = "matched_geneIDs.gtf", clusters = 4)
}
\arguments{
\item{query}{Assembled transcripts in query. Argument can be provided as name of GTF/GFF3 file in current working directory,
or path_to_file. Argument can also be provided as a GenomicRanges object containing transcript and/or exon ranges.}

\item{ref}{Annotated reference transcripts. This program contain gencode basic annotations from mm10 and hg38 assemblies
that will be loaded using 'gencode_basics[['mm10']] or gencode_basics[['hg38']] as input to argument. Argument can also be provided as name of 
GTF/GFF3 file in current working directory, or path_to_file.}

\item{primary_gene_id}{Name of the primary gene id in query file. Input to this argument is typically 'gene_id'}

\item{secondary_gene_id}{Name of the secondary gene id in query file. Example of input to this arguement is 'ref_gene_id'}

\item{outputfile}{Name of output file (Default: 'matched_geneIDs.gtf'). File will be saved in current working directory}

\item{query_format}{Format of query file (GTF/GFF3). If this argument is not provided and input to query argument
is a filename, program will attempt to read file extension.}

\item{ref_format}{Format of reference file (GTF/GFF3). If this argument is not provided and input to query argument
is a filename, program will attempt to read file extension.}
}
\value{
GTF file with matched Gene IDs if makefile == TRUE. 
GRanges object if makefile == FALSE.
}
\description{
This function will match and correct Gene IDs from a query assembled transcript file, using a 
transcript annotation as reference. 

The default approach to this correction relies on finding overlaps between transcripts in query with
transcripts in reference. Using this method alone could result in false positive matches (19 percent false positives).
To improve this, users have an option to activate two additional layers of matching.
(1) Matching by ENSEMBL Gene_IDs. If both query and reference transcript annotations containg Ensembl-style
Gene IDs, this program will try to match both IDs in a less stringent manner. This correction can be invoked
by providing the 'primary_gene_id' argument

(2) Matching by secondary Gene_IDs. Depending on the transcript assembly program, GTF/GFF3 annotations 
may contain additional comments on the transcript information. This may include a distinct 
secondary Gene ID annotation that potentially matches with the reference. To invoke this correction,
provide 'primary_gene_id' and 'secondary_gene_id' arguments. To determine if your transcript assembly contain 
possible secondary Gene IDs, try importing query GTF file using rtracklayer package and check its metadata 
columns
}
\examples{

matchGeneIDs(testData, gencode_basics[['mm10']], primary_gene_id = 'gene_id', secondary_gene_id = 'ref_gene_id', clusters = 4)



}
