% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSeqLevels.R
\name{matchSeqLevels}
\alias{matchSeqLevels}
\title{Match seqnames of input GRanges to reference GRanges}
\usage{
matchSeqLevels(inGRanges, refGRanges)
}
\arguments{
\item{inGRanges}{GRanges object with seqnames to change}

\item{refGRanges}{GRanges object from which seqnames is referenced}
}
\value{
Corrected input GRanges
}
\description{
Match seqnames of input GRanges to reference GRanges
}
