% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_compare.R
\name{plot.auto_compare}
\alias{plot.auto_compare}
\title{Plot method for auto_compare objects}
\usage{
\method{plot}{auto_compare}(x, ...)
}
\arguments{
\item{x}{An object of class "auto_compare".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot method for auto_compare objects
}
