% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCDS.R
\name{predictCDS}
\alias{predictCDS}
\title{Title}
\usage{
predictCDS(query, refCDS, fasta, query2ref, ids = c(1, 2), coverage = NULL)
}
\arguments{
\item{query}{}

\item{refCDS}{}

\item{fasta}{}

\item{query2ref}{}

\item{ids}{}

\item{coverage}{}
}
\value{

}
\description{
Title
}
\examples{
predictCDS(querytx, refCDS, Mmusculus, q2r_df)

}
