% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMD.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict NMD nature of mRNA transcripts}
\usage{
predictNMD(tx, cds, NMDthreshold = 50, which = NULL, return = c("NMD", "all"))
}
\arguments{
\item{tx}{GRanges object or GRangesList object containing exons
for each transcript}

\item{cds}{GRanges object or GRangesList object containing coding regions (CDS)
for each transcript. GRangesList must have names that match names in tx,
else tx with missing cds will not be analysed}

\item{NMDthreshold}{Minimum distance of STOP codon to last exon junction (EJ) which triggers NMD.
Default = 50bp}

\item{which}{List containing tx names to filter for analysis}

\item{return}{If tx and cds are GRangesList, returns only NMD-sensitive
transcripts or all transcripts. Default = 'NMD'}
}
\value{
List with prediction of NMD sensitivity and statistics:

is_NMD: logical value in prediciting transcript sensitivity to NMD

dist_to_lastEJ: Integer value indicating distance of STOP codon to last EJ
Values are referenced from last EJ, thus a positive value indicates upstream
position of STOP codon while negative value indicates downstream position

num_of_down_EJs: Number of downstream EJs

dist_to_downEJs: Integer value indicating distance of STOP codon to each down 
EJs Values are referenced from last EJ, thus a positive value indicates 
upstream position of STOP codon while negative value indicates downstream 
position together with distances of STOP codon to last EJ
}
\description{
Predicts NMD-sensitivity of transcripts
}
\examples{

### To visualize transcripts
library(wiggleplotr)
plotTranscripts(txl[1:2], cdsl[1:2])
plotTranscripts(txl[3:4], cdsl[3:4])

### Examples with single GRanges objects
predictNMD(txl$ENSMUST00000029780,cdsl$ENSMUST00000029780) # NMD-insensitive
predictNMD(txl$ENSMUST00000197833,cdsl$ENSMUST00000197833) # NMD-sensitive
###

### Examples with GRangesList object
predictNMD(txl,cdsl)
predictNMD(txl,cdsl, return = 'all')
predictNMD(txl,cdsl,which=c('ENSMUST00000029780', 'ENSMUST00000197833'), return = 'all')



}
\author{
Fursham Hamid
}
