% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.prepareAnalysis.R
\name{prepareAnalysis}
\alias{prepareAnalysis}
\title{NMDer workflow: Preparing for analysis}
\usage{
prepareAnalysis(inputGRanges, basicGRanges)
}
\arguments{
\item{inputGRanges}{Query GRanges object}

\item{basicGRanges}{Reference GRanges object.}
}
\value{
Function will return a list containing:
(1) dataframe of a list of query transcripts for downstream workflow
(2) GRangesList of exon coordinates from query as dataframe
(3) GRangesList of exon coordinates from reference as dataframe
(4) GRangesList of exon coordinates of reference CDS as dataframe
}
\description{
THIS FUNCTION IS PART OF THE NMDer PROGRAM WORKFLOW.
This function will prepare output dataframes and GRanges objects for downstream workflow
}
