% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepNMDer}
\alias{prepNMDer}
\title{Title}
\usage{
prepNMDer(query, reference, fasta, query_format = NULL,
  reference_format = NULL, match_chrom = FALSE, match_geneIDs = FALSE,
  primary_gene_id = NULL, secondary_gene_id = NULL, clusters = 4)
}
\arguments{
\item{query}{}

\item{reference}{}

\item{fasta}{}

\item{query_format}{}

\item{reference_format}{}

\item{match_chrom}{}

\item{match_geneIDs}{}

\item{primary_gene_id}{}

\item{secondary_gene_id}{}

\item{clusters}{}
}
\description{
Title
}
