% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{prePonder}
\alias{prePonder}
\title{PONDER workflow: Prepare PONDER analysis}
\usage{
prePonder(
  query,
  reference,
  fasta,
  query_format = NULL,
  reference_format = NULL,
  match_chrom = FALSE,
  match_geneIDs = FALSE,
  primary_gene_id = NULL,
  secondary_gene_id = NULL
)
}
\arguments{
\item{query}{Mandatory. Path to query GTF/GFF3 transcript annotation file}

\item{reference}{Mandatory. Path to reference GTF/GFF3 transcript annotation file.}

\item{fasta}{Mandatory. BSGenome object (preferred) or path to fasta file}

\item{query_format}{Optional argument to specify the query annotation format ('gtf','gff3'). 
Mandatory if query contains '.txt' extension filename}

\item{reference_format}{Optional argument to specify the reference annotation format ('gtf','gff3'). 
Mandatory if reference contains '.txt' extension filename}

\item{match_chrom}{Supplementary feature. If TRUE, program will attempt to match chromosome names of 
query and reference to fasta genome to ensure consistent naming across input files.}

\item{match_geneIDs}{Supplementary feature to attempt to match gene IDs in query file
to reference file. This is key in grouping query transcripts to reference gene families for comparison.


Matching is done at three levels with increasing accuracy:

1. Crudely intersecting query coordinates with reference. Invoked by setting match_geneIDs to TRUE

2. Trim ensembl-style gene IDs and attempt matching. Invoked by providing name of 
gene ID header (typically 'gene_id') from gtf file to primary_gene_id argument

3. Replace query gene ID with a secondary gene ID and attempt matching. Invoked by providing name of 
secondary gene ID header (for example 'ref_gene_id') from gtf file to secondary_gene_id argument}

\item{primary_gene_id}{See match_geneIDs argument}

\item{secondary_gene_id}{See match_geneIDs argument}
}
\value{
S4 object containing dataframes and objects for downstream NMD prediction analysis
}
\description{
Import transcript annotation file, 
match chromosome levels and gene IDs and prepare for NMD prediction analysis
}
