% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.preTesting.R
\name{preTesting}
\alias{preTesting}
\title{NMDer workflow: Check and match input annotations}
\usage{
preTesting(inputGRanges, basicGRanges, genome, correct_chrom,
  correct_gene_id, primary_gene_id, secondary_gene_id)
}
\arguments{
\item{inputGRanges}{Query GRanges object}

\item{basicGRanges}{Reference GRanges object}

\item{genome}{Genome sequence as a Biostring object}

\item{correct_chrom}{Supplementary feature. If TRUE, program will attempt to match chromosome names of 
query and reference to fasta genome to ensure consistent naming across input files.}

\item{correct_gene_id}{Supplementary feature to attempt to match gene IDs in query file
to reference file. This is key in grouping query transcripts to reference gene families for comparison.


Matching is done at three levels with increasing accuracy:

1. Crudely intersecting query coordinates with reference. Invoked by setting match_geneIDs to TRUE
sprintf('Remaining number of non-standard gene_ids: %s', nonstand_after)
2. Trim ensembl-style gene IDs and attempt matching. Invoked by providing name of 
gene ID header (typically 'gene_id') from gtf file to primary_gene_id argument

3. Replace query gene ID with a secondary gene ID and attempt matching. Invoked by providing name of 
secondary gene ID header (for example 'ref_gene_id') from gtf file to secondary_gene_id argument}

\item{primary_gene_id}{See 'correct_gene_id' for details}

\item{secondary_gene_id}{See 'correct_gene_id' for details}
}
\value{
Query GRanges object that have been matched
}
\description{
This program checks the chromosome ID and gene ID of query, 
reference and genome objects and if requested, attempts to match these entries
}
