% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_compare.R
\name{print.auto_compare}
\alias{print.auto_compare}
\title{Print method for auto_compare objects}
\usage{
\method{print}{auto_compare}(x, ...)
}
\arguments{
\item{x}{An object of class "auto_compare".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for auto_compare objects
}
