% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_compare_cat.R
\name{print.auto_compare_cat}
\alias{print.auto_compare_cat}
\title{Print method for auto_compare_cat objects}
\usage{
\method{print}{auto_compare_cat}(x, ...)
}
\arguments{
\item{x}{An object of class "auto_compare_cat".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for auto_compare_cat objects
}
