% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{print.plot_volcano}
\alias{print.plot_volcano}
\title{Print Method for \code{plot_volcano} Objects}
\usage{
\method{print}{plot_volcano}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"plot_volcano"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Compact console summary of a \code{plot_volcano} result.
}
\author{
John Lennon L. Calorio
}
