% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assoc.R
\name{print.run_assoc}
\alias{print.run_assoc}
\title{Print method for run_assoc objects}
\usage{
\method{print}{run_assoc}(x, ...)
}
\arguments{
\item{x}{An object of class "run_assoc".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for run_assoc objects
}
