% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assoc.R
\name{print.run_assoc_error}
\alias{print.run_assoc_error}
\title{Print method for run_assoc_error objects}
\usage{
\method{print}{run_assoc_error}(x, ...)
}
\arguments{
\item{x}{An object of class \code{run_assoc_error}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for run_assoc_error objects
}
