% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_auc.R
\name{print.run_auc}
\alias{print.run_auc}
\title{Print Method for \code{run_auc} Objects}
\usage{
\method{print}{run_auc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"run_auc"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Compact summary of a \code{run_auc} result.
}
\author{
John Lennon L. Calorio
}
