% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_foldchange.R
\name{print.run_foldchange}
\alias{print.run_foldchange}
\title{Print Method for \code{run_foldchange} Objects}
\usage{
\method{print}{run_foldchange}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"run_foldchange"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Compact console summary of a \code{run_foldchange} result.
}
\author{
John Lennon L. Calorio
}
