% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_reduce.R
\name{print.run_reduce}
\alias{print.run_reduce}
\title{Print Method for \code{run_reduce} Objects}
\usage{
\method{print}{run_reduce}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"run_reduce"}.}

\item{...}{Further arguments passed to or from other methods (currently
unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Provides a concise, informative summary when a
\code{run_reduce} object is printed to the console.
}
\author{
John Lennon L. Calorio
}
