% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{reconstructCDS}
\alias{reconstructCDS}
\title{Reconstruct ORF with alternative segments}
\usage{
reconstructCDS(knownCDS, queryTx, txrevise_out, annotatedStart = TRUE)
}
\arguments{
\item{knownCDS}{A GRanges object containing CCDS information of a gene family}

\item{queryTx}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{txrevise_out}{Optional. 
GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information.
Arguments knownCDS and queryTx is not mandatory if this argument is provided}

\item{annotatedStart}{A}
}
\value{
A GRanges Object with new ORF information if transcript contain unique segments, or
NA if transcript is identical to reference CDS
}
\description{
This function will generate a new ORF of a gene family with 
added segments from a query transcript. INPUT...
}
\examples{

reconstructCDS(_)





}
\author{
Fursham Hamid
}
