% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{reconstructCDSstart}
\alias{reconstructCDSstart}
\title{Reconstruct CDS with alternate 5' exons}
\usage{
reconstructCDSstart(knownCDS, queryTx, txrevise_out, refsequence)
}
\arguments{
\item{knownCDS}{A GRanges object containing CCDS information of a gene family}

\item{queryTx}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{txrevise_out}{GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information}

\item{refsequence}{sequence file build from AnnotationHub}
}
\value{
A list containing (1) a GRanges object with a reconstructed CDS, 
(2) GRangesobject from 'indentifyAddedRemovedRegions()' and 
(3) a flag for non-annotated Start codon.
If reconstructed CDS do not contain an in-frame Start codon with 3' end of the coding sequence,
function will return NA for (2)
}
\description{
This function will reconstruct a CDS sequence with a different 5' end 
taken from a query transcript from the same gene
}
\examples{

reconstructCDSstart(ptbp2_testData$noNMD, ptbp2_testData$diffstart, refsequence = mmus_dna)

}
\author{
Fursham Hamid
}
