% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNMD_functions.R
\name{resizeTranscripts}
\alias{resizeTranscripts}
\title{Resize GRanges transcript object}
\usage{
resizeTranscripts(x, start = 0, end = 0)
}
\arguments{
\item{x}{GRanges object containing exon coordinates of a transcript or CDS}

\item{start}{Length of 5' end to truncate (positive val) or extend (negative val)}

\item{end}{Length of 3' end to truncate (positive val) or extend (negative val)}
}
\value{
a new GRanges transcript object
}
\description{
This function will append/extend the 5' and 3' ends of a GRanges object which describes the 
exon ranges of a transcript or a CDS
}
\examples{
resizeTranscripts(ptbp2_testData$exons$ENSMUST00000197833, start = 100)

}
