% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run}
\alias{run}
\title{Run _ Program}
\usage{
run(file, reference, fasta, primary_gene_id = NULL,
  secondary_gene_id = NULL, nonclassical = FALSE, makegtf = FALSE,
  PTC_dist = 50, output_dir = "NMDer", quiet = FALSE)
}
\arguments{
\item{file}{I}

\item{reference}{}

\item{fasta}{}

\item{primary_gene_id}{}

\item{secondary_gene_id}{}

\item{nonclassical}{}

\item{makegtf}{}

\item{PTC_dist}{}

\item{quiet}{}
}
\description{

}
