% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runNMDer}
\alias{runNMDer}
\title{Title}
\usage{
runNMDer(prepObject, testNMD = TRUE, other_features = FALSE,
  PTC_dist = 50, testAS = FALSE, clusters = 4)
}
\arguments{
\item{prepObject}{}

\item{testNMD}{}

\item{other_features}{}

\item{PTC_dist}{}

\item{testAS}{}

\item{clusters}{}
}
\description{
Title
}
