% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_select.R
\name{select_numvars}
\alias{select_numvars}
\title{Select Numeric Variables from a Data Frame}
\usage{
select_numvars(
  x,
  force_num = NULL,
  include_int = TRUE,
  include_dbl = TRUE,
  include_lgl = TRUE,
  include_comp = TRUE
)
}
\arguments{
\item{x}{A data frame or matrix.}

\item{force_num}{Character vector of column names in \code{x} to be
forcefully included as numeric variables.}

\item{include_int}{Logical. If \code{TRUE} (default), includes integer
type variables.}

\item{include_dbl}{Logical. If \code{TRUE} (default), includes double
type variables.}

\item{include_lgl}{Logical. If \code{TRUE} (default), includes logical
type variables.}

\item{include_comp}{Logical. If \code{TRUE} (default), includes complex
type variables.}
}
\value{
A data frame containing only the selected numeric variables.
}
\description{
This function selects variables from a data frame based on their numeric
types (integer, double, logical, or complex). It also allows forcing
specific columns to be treated as numeric even if they are stored as
other types.
}
\examples{
df <- data.frame(a = 1:3, b = c(1.1, 2.2, 3.3), 
c = c("x", "y", "z"),
d = c(TRUE, FALSE, TRUE))
select_numvars(df)
}
\author{
John Lennon L. Calorio
}
