% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_compare.R
\name{summary.auto_compare}
\alias{summary.auto_compare}
\title{Summary method for auto_compare objects}
\usage{
\method{summary}{auto_compare}(object, ...)
}
\arguments{
\item{object}{An object of class "auto_compare".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for auto_compare objects
}
