% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_compare_cat.R
\name{summary.auto_compare_cat}
\alias{summary.auto_compare_cat}
\title{Summary method for auto_compare_cat objects}
\usage{
\method{summary}{auto_compare_cat}(object, ...)
}
\arguments{
\item{object}{An object of class "auto_compare_cat".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for auto_compare_cat objects
}
