% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assoc.R
\name{summary.run_assoc}
\alias{summary.run_assoc}
\title{Summary method for run_assoc objects}
\usage{
\method{summary}{run_assoc}(object, ...)
}
\arguments{
\item{object}{An object of class "run_assoc".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for run_assoc objects
}
