% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_assoc.R
\name{summary.run_assoc_multi}
\alias{summary.run_assoc_multi}
\title{Summary method for run_assoc_multi objects}
\usage{
\method{summary}{run_assoc_multi}(object, ...)
}
\arguments{
\item{object}{An object of class \code{run_assoc_multi} from \code{run_assoc}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for run_assoc_multi objects
}
