% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diff.R
\name{summary.run_diff}
\alias{summary.run_diff}
\title{Summary method for run_diff objects}
\usage{
\method{summary}{run_diff}(object, ...)
}
\arguments{
\item{object}{An object of class "run_diff".}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary method for run_diff objects
}
