% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNMD.R
\name{testNMD}
\alias{testNMD}
\title{Test ORF for NMD}
\usage{
testNMD(queryCDS, queryTranscript, distance_stop_EJ = 50,
  other_features = FALSE, fasta)
}
\arguments{
\item{queryCDS}{GRanges Object containing CDS structure of a transcript}

\item{queryTranscript}{GRanges Object containing transcript structure of the same transcript}

\item{distance_stop_EJ}{NMD Threshold. Distance of stop codon to last exon-exon junction. Default: 50}

\item{other_features}{TRUE/FALSE argument to report other NMD features. Default: FALSE}

\item{fasta}{Fasta sequence file. Mandatory if other_features == TRUE}
}
\value{
A list containing information on: 
(1) NMD nature of transcript
(2) Distance of stop codon to last exon-exon junction
if other features of NMD is to be returned:
(3) coordinates of uORF
(4) coordinates of uATG
(5) frame of uATG
}
\description{
This function will test for NMD features of a transcript with a given open reading frame
}
\examples{

library("BSgenome.Mmusculus.UCSC.mm10")
testNMD(ptbp2Data$refCDS, ptbp2Data$transcripts$ENSMUST00000029780)
testNMD(ptbp2Data$refCDS, ptbp2Data$transcripts$ENSMUST00000029780, other_features = TRUE, fasta = BSgenome.Mmusculus.UCSC.mm10)

testNMD(ptbp2Data$skipE10CDS, ptbp2Data$transcripts$ENSMUST00000197833)
testNMD(ptbp2Data$skipE10CDS, ptbp2Data$transcripts$ENSMUST00000197833, other_features = TRUE, fasta = BSgenome.Mmusculus.UCSC.mm10)


}
\author{
Fursham Hamid
}
