% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNMD_.R
\name{testNMD_}
\alias{testNMD_}
\title{Test ORF for NMD}
\usage{
testNMD_(
  queryTranscript,
  queryCDS,
  distance_stop_EJ = 50,
  other_features = FALSE,
  fasta
)
}
\arguments{
\item{queryTranscript}{GRanges Object containing transcript structure of the same transcript}

\item{queryCDS}{GRanges Object containing CDS structure of a transcript}

\item{distance_stop_EJ}{NMD Threshold. Distance of stop codon to last exon-exon junction. Default: 50}

\item{other_features}{TRUE/FALSE argument to report other NMD features. Default: FALSE}

\item{fasta}{Fasta sequence file. Mandatory if other_features == TRUE}
}
\value{
A list containing information on: 
(1) NMD nature of transcript
(2) Distance of stop codon to last exon-exon junction
if other features of NMD is to be returned:
(3) coordinates of uORF
(4) coordinates of uATG
(5) frame of uATG
}
\description{
This function will test for NMD features of a transcript with a given open reading frame
}
\author{
Fursham Hamid
}
