% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{testNMDfeatures}
\alias{testNMDfeatures}
\title{_ workflow: Test transcripts for NMD feature}
\usage{
testNMDfeatures(report_df, inputExonsbyTx, basicExonsbyCDS, basicExonsbyTx,
  genome, PTC_dist = 50, testNonClassicalNMD = FALSE)
}
\arguments{
\item{report_df}{dataframe of a list of assembled transcripts. It is recommended to create this object using
prepareAnalysis function from the workflow. see ?prepareAnalysis}

\item{inputExonsbyTx}{GRangesList of exon coordinates of assembled transcripts grouped by transcript name. It is recommended to create this object using
prepareAnalysis function from the workflow. see ?prepareAnalysis}

\item{basicExonsbyCDS}{GRangesList of exon coordinates of reference CDS grouped by transcript name. It is recommended to create this object using
prepareAnalysis function from the workflow. see ?prepareAnalysis}

\item{genome}{DNAstring containing genome sequence}

\item{PTC_dist}{Numerical value referring to minimium distance of premature stop codon to last exon junction}

\item{testNonClassicalNMD}{Boolean value (TRUE/FALSE) on whether to test for non-classical NMD features}

\item{basicTX_df}{dataframe of a list of reference CDS transcripts. It is recommended to create this object using
prepareAnalysis function from the workflow. see ?prepareAnalysis}
}
\value{
updated form of report_df
}
\description{
THIS FUNCTION IS PART OF THE _ PROGRAM WORKFLOW.
This function will analyze the assembled transcripts for NMD features
}
