% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{testNMDtx}
\alias{testNMDtx}
\title{This function will test if the CDS in query is an NMD substrate}
\usage{
testNMDtx(queryCDS, refsequence)
}
\arguments{
\item{queryCDS}{is a GRanges Object}

\item{refsequence}{is a sequence containing TwoBitFile}
}
\value{
Granges
}
\description{
This function will test if the CDS in query is an NMD substrate
}
