% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{testNMDvsThisCDS}
\alias{testNMDvsThisCDS}
\title{_ workflow: Test transcript for NMD feature against a reference CDS}
\usage{
testNMDvsThisCDS(knownCDS, queryTx, refsequence, PTC_dist = 50,
  nonClassicalNMD = FALSE)
}
\arguments{
\item{knownCDS}{}

\item{queryTx}{}

\item{refsequence}{}

\item{PTC_dist}{}

\item{nonClassicalNMD}{}
}
\description{
THIS FUNCTION IS PART OF THE _ PROGRAM WORKFLOW.
This function will compare the query transcript with a reference CDS and tests
if insertion of alternative segments into the CDS generates an NMD substrate
}
