% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNMD_functions.R
\name{testNonClassicalNMD}
\alias{testNonClassicalNMD}
\title{Title}
\usage{
testNonClassicalNMD(testCDS, testTranscript, fasta)
}
\arguments{
\item{testCDS}{}

\item{testTranscript}{}

\item{fasta}{}
}
\description{
Title
}
