% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{testTXforNMD}
\alias{testTXforNMD}
\title{Test ORF for NMD}
\usage{
testTXforNMD(queryCDS, refsequence, minmPTCdist = -50, full.output = FALSE)
}
\arguments{
\item{queryCDS}{GRanges Object containing CDS structure}

\item{refsequence}{sequence file build from AnnotationHub}

\item{full.output}{If TRUE, function will output other information}
}
\value{
A list containing: 
(1) GRanges object containing a list of in-frame stop codons and its metadata

if full.output == TRUE, function will also return
(2) An ORF which ends at the most upstream in-frame stopcodon
}
\description{
This function will check for NMD features for an input ORF

Do not that input query has to begin with an ATG
}
\examples{

testTXforNMD(ptbp2_testData$noNMD, mmus_dna)
testTXforNMD(ptbp2_testData$NMD, mmus_dna)

}
\author{
Fursham Hamid
}
