% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_modules.R
\name{testTXforStart}
\alias{testTXforStart}
\title{Test transcripts for annotated start codons}
\usage{
testTXforStart(refCDS, queryTranscript, full.output = FALSE)
}
\arguments{
\item{queryTranscript}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{refCDS}{A GRanges object containing CCDS information of a gene family}

\item{full.output}{If TRUE, function will output additional information}
}
\value{
By default, function will return a TRUE/FALSE boolean (annotatedStart)

If full.output is TRUE, function will output other information which include:
(1) the length of the first coding exon (firstexlen), 
(2) queryTranscript GRanges with 5' end appended to the start codon  (ORF),  
(3) a list of GRanges objects similar to 'indentifyAddedRemovedRegions()' output (txrevise_out). 
?indentifyAddedRemovedRegions for more information
}
\description{
This function will test whether the query transcript contain an annotated 
Start codon from a reference CCDS of the gene
}
\examples{

testTXforStart(ptbp2Data$transcripts$ENSMUST00000197833, ptbp2Data$refCDS)
testTXforStart(ptbp2Data$afCDS, ptbp2Data$refCDS)


}
\author{
Fursham Hamid
}
