% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_correl.R
\name{as_gt}
\alias{as_gt}
\alias{as_gt.run_correl}
\title{Convert a run_correl to a gt table}
\usage{
as_gt(x, ...)

\method{as_gt}{run_correl}(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments.}
}
\value{
A \code{gt_tbl} object.
Convert an object to a gt table

A gt table object.
}
\description{
Passes the data frame to \code{gt::gt()} and appends the method footnote
(if present) as a source note via \code{gt::tab_source_note()}.
}
