% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.prepareInputs.R
\name{prepareInputs}
\alias{prepareInputs}
\title{NMDer workflow: Import files}
\usage{
prepareInputs(queryFile, refFile, fasta, user_query_format, user_ref_format)
}
\arguments{
\item{queryFile}{Path to query GTF/GFF3 transcript annotation file}

\item{refFile}{Path to ref GTF/GFF3 transcript annotation file}

\item{fasta}{BSGenome object (preferred) or path to fasta file}

\item{user_query_format}{optional argument to specify the query annotation format ('gtf','gff3')}

\item{user_ref_format}{optional argument to specify the refFileerence annotation format ('gtf','gff3')}
}
\value{
A list containing (1) GRanges object for query, (2) GRanges object for 
ref annotation, and (3) DNAstring containing genome sequence
}
\description{
Imports mandatory files for NMDer analysis
}
