% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunctions.R
\name{prepareInputs}
\alias{prepareInputs}
\title{_ workflow: Load and/or import input files}
\usage{
prepareInputs(queryfile, ref, fasta = NULL, in_format, ref_format)
}
\arguments{
\item{fasta}{genome sequence in fasta format. This program contain in-built fasta files for 
GRCh38_hg38 and GRCm38_mm10 genome. These sequences can be loaded by parsing "hg38" or "mm10" 
into this argument. If users wish to import other genomes, please provide filename or 
/path/to/file of fasta files.}

\item{file}{filename or /path/to/file of transcript assemblies. File can be of
gtf, gff3 or bed format, and its filename should contain the corresponding extension}

\item{gencode}{reference annotation of coding transcripts for each gene family. This program
contain in-built gencode_basic annotation for GRCh38_hg38 and GRCm38_mm10 assemblies. These
assemblies can be loaded by parsing "hg38" or "mm10" into this argument. If users wish to 
import other assemblies, please provide filename or /path/to/file of reference CDS assemblies.
File can be of gtf, gff3 or bed format, and its filename should contain the corresponding extension}
}
\value{
A list containing (1) GRanges object for assembled transcripts, (2) GRanges object for 
reference CDS assembly, and (3) DNAstring containing genome sequence
}
\description{
THIS FUNCTION IS PART OF THE _ PROGRAM WORKFLOW. 
This function will import and/or load input data for the workflow and
will return a list containing GRanges objects for assembled transcript and 
reference annotation, and fasta file for genome sequence
}
