% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunctions.R
\name{preTesting}
\alias{preTesting}
\title{_ workflow: Testing and correcting transcript assembly}
\usage{
preTesting(inputGRanges, basicGRanges, genome, correct_chrom,
  correct_gene_id, primary_gene_id, secondary_gene_id)
}
\arguments{
\item{inputGRanges}{GRanges object containing transcript and exon features. It is recommended to create this object
by importing gtf/gff3 files using rtracklayer::import}

\item{basicGRanges}{#' GRanges object containing reference transcript,exon and CDS features. 
It is recommended to create this object by importing gtf/gff3 files using rtracklayer::import}

\item{genome}{DNAstring object containing genome sequence}

\item{primary_gene_id}{Optional argument. Typically, this is 'gene_id'. Refer to description for details}

\item{secondary_gene_id}{Optional argument. Refer to description for details}
}
\value{
Function will return a modified inputGRanges object
}
\description{
THIS FUNCTION IS PART OF THE _ PROGRAM WORKFLOW. 
This function will test and correct the input transcript assembly for consistent chromosome names 
and gene ids with CDS assembly and genome sequence.

Non-consistent Gene IDs are corrected by replacing it with Gene IDs from reference CDS assembly
that contain overlapping exons. However, this strategy may be inaccurate and may return 
false-postive gene ids. To improve the correction, we provide two upper layers of correction
that can be initiated by providing the mandatory 'primary_gene_id' and 'secondary_gene_id'
arguments

Primary_gene_id refer to the name of the metadata column containing gene ids for each assemnbled transcript. 
This is typically 'gene_id'
Secondary_gene_id refer to the name of the metadata column that may contain a reference gene id
For example: 
NEED TO DESCRIBE
 
But if both args are provided, 
function will try to replace 

How to check for metadata columns
}
