% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.preTesting.R
\name{preTesting}
\alias{preTesting}
\title{NMDer workflow: Check and match input annotations}
\usage{
preTesting(inputGRanges, basicGRanges, genome, correct_chrom,
  correct_gene_id, primary_gene_id, secondary_gene_id)
}
\arguments{
\item{inputGRanges}{Query GRanges object}

\item{basicGRanges}{Reference GRanges object}

\item{genome}{Genome sequence as a Biostring object}

\item{correct_chrom}{If TRUE, attempt to match chromosome ID}

\item{correct_gene_id}{If TRUE, attempt to match gene ID. This is crudely done by intersecting query transcript coordinates with reference. To increase chance of true matching, user may provide}

\item{primary_gene_id}{See 'correct_gene_id' for details}

\item{secondary_gene_id}{See 'correct_gene_id' for details}
}
\value{
Query GRanges object that have been matched
}
\description{
Check the chromosome ID and gene ID of query, 
reference and genome objects and if requested, attempts to match these metadata
}
