% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNMD_functions.R
\name{reconstructCDS}
\alias{reconstructCDS}
\title{Reconstruct ORF with alternative internal and downstream segments}
\usage{
reconstructCDS(knownCDS, queryTx, txrevise_out)
}
\arguments{
\item{knownCDS}{A GRanges object containing CCDS information of a gene family}

\item{queryTx}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{txrevise_out}{Optional. 
GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information.
Arguments knownCDS and queryTx is not mandatory if this argument is provided}
}
\value{
A GRanges Object with new ORF information if transcript contain unique segments, or
NA if transcript is identical to reference CDS
}
\description{
This function will generate a new ORF of a gene family with 
added segments from a query transcript. 

Do note that this function, if ran as a stand-alone, 
will not insert/remove first exons.
}
\examples{

reconstructCDS(ptbp2_testData$noNMD, ptbp2_testData$exons$ENSMUST00000197833)





}
\author{
Fursham Hamid
}
