% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_reconstructCDS.R
\name{reconstructCDS}
\alias{reconstructCDS}
\title{Reconstruct CDS with alternative internal and downstream segments}
\usage{
reconstructCDS(
  queryTranscript,
  refCDS,
  fasta,
  txrevise_out = NULL,
  gene_id,
  transcript_id
)
}
\arguments{
\item{queryTranscript}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{refCDS}{A GRanges object containing CCDS information of a gene family}

\item{fasta}{Fasta sequence of the genome}

\item{txrevise_out}{Optional. 
GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information.
Arguments refCDS and queryTranscript is not mandatory if this argument is provided}
}
\value{
A list containing:
(1) A GRanges Object of new ORF, or NA if no ORF is found
(2) TRUE/FALSE object on whether ORF is an alternative CDS transcript
}
\description{
This function will add alternative segments from a query transcript into 
a reference CDS from the same gene and generate a new ORF

Note: This function will not insert/remove first exons.
}
\examples{

library("BSgenome.Mmusculus.UCSC.mm10")
reconstructCDS(ptbp2Data$transcripts$ENSMUST00000197833, ptbp2Data$refCDS, fasta = BSgenome.Mmusculus.UCSC.mm10)


}
\author{
Fursham Hamid
}
