% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_modules.R
\name{reconstructCDSstart}
\alias{reconstructCDSstart}
\title{Reconstruct CDS with alternate 5' exons}
\usage{
function(queryTranscript, refCDS, txrevise_out, fasta, full.output = FALSE)
}
\arguments{
\item{queryTranscript}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{refCDS}{A GRanges object containing reference CDS information of a gene family}

\item{fasta}{Fasta sequence file}

\item{txrevise_out}{Optional input. GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information}

\item{full.output}{if TRUE, function will output additional information}
}
\value{
Default output (ORF_considered),
If reconstructed CDS contain an in-frame ATG, function will output a GRanges Object describing the
reconstructed CDS. Else, function will output 'NA'

If full.output == TRUE,
function will also return a list of GRanges objects similar to 'indentifyAddedRemovedRegions()' output (txrevise_out). 
?indentifyAddedRemovedRegions for more information
}
\description{
This function will reconstruct a CDS sequence with a 5' end 
taken from a query transcript from the same gene
}
\examples{

reconstructCDSstart(ptbp2Data$afCDS, ptbp2Data$refCDS, fasta = Ensembl_mm10)

}
\author{
Fursham Hamid
}
