% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{reconstructCDSstart}
\alias{reconstructCDSstart}
\title{Reconstruct CDS with alternate 5' exons}
\usage{
reconstructCDSstart(knownCDS, queryTx, txrevise_out, refsequence, full.output = FALSE)
}
\arguments{
\item{knownCDS}{A GRanges object containing CCDS information of a gene family}

\item{queryTx}{A GRanges object containing exon structure of a query transcript from
the same gene family}

\item{txrevise_out}{GRangesobject from 'indentifyAddedRemovedRegions()' 
or 'testTXforStart()' output. ?indentifyAddedRemovedRegions or ?testTXforStart
for more information}

\item{refsequence}{sequence file build from AnnotationHub}

\item{full.output}{if TRUE, function will output additional information}
}
\value{
Default output (ORF),
If reconstructed CDS contain an in-frame ATG, function will output a GRanges Object describing the
reconstructed CDS. Else, function will output 'NA'

If full.output == TRUE,
function will also return a list of GRanges objects similar to 'indentifyAddedRemovedRegions()' output (txrevise_out). 
?indentifyAddedRemovedRegions for more information
}
\description{
This function will reconstruct a CDS sequence with a different 5' end 
taken from a query transcript from the same gene
}
\examples{

reconstructCDSstart(ptbp2_testData$noNMD, ptbp2_testData$diffstart, refsequence = mmus_dna)

}
\author{
Fursham Hamid
}
