% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execs.R
\name{runNMDer}
\alias{runNMDer}
\title{NMDer workflow: Run NMDer analysis}
\usage{
runNMDer(
  prepObject,
  testNMD = TRUE,
  testOtherFeatures = FALSE,
  PTC_dist = 50,
  testAS = FALSE,
  makeGTF = FALSE,
  clusters = 1
)
}
\arguments{
\item{prepObject}{S4 object from prepNMDer output}

\item{testNMD}{If TRUE, program will test transcripts for primary NMD feature}

\item{testOtherFeatures}{If TRUE, program will test transcripts for other NMD features}

\item{PTC_dist}{Minimum distance of PTC to last exon-junction to trigger NMD.
Default: 50bp}

\item{testAS}{If TRUE, program will test for alternatively spliced segments}

\item{makeGTF}{If TRUE, program will output a GTF file of query
Provide path to output}

\item{clusters}{Number of cores to run the program. Default = 4}
}
\value{
df with analysis
}
\description{
Execute core analysis
}
\examples{

}
