% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execs.R
\name{runNMDer}
\alias{runNMDer}
\title{Title}
\usage{
runNMDer(prepObject, testNMD = TRUE, testOtherFeatures = FALSE,
  PTC_dist = 50, testAS = FALSE, makeGTF = FALSE, clusters = 4)
}
\arguments{
\item{prepObject}{}

\item{testNMD}{}

\item{PTC_dist}{}

\item{testAS}{}

\item{clusters}{}

\item{other_features}{}
}
\description{
Title
}
