% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{testTxforNMD}
\alias{testTxforNMD}
\title{Test query CDS for NMD features}
\usage{
testTxforNMD(queryCDS, refsequence)
}
\arguments{
\item{queryCDS}{GRanges Object containing CDS structure}

\item{refsequence}{sequence file build from AnnotationHub}
}
\value{
A GRanges object containing a list of in-frame stop codons, its
coordinate, its distance to last EJC and prediction of its NMD nature
}
\description{
Test query CDS for NMD features
}
\author{
Fursham Hamid
}
