% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{testTXforNMD}
\alias{testTXforNMD}
\title{Test ORF for NMD}
\usage{
testTXforNMD(queryCDS, refsequence)
}
\arguments{
\item{queryCDS}{GRanges Object containing CDS structure}

\item{refsequence}{sequence file build from AnnotationHub}
}
\value{
A GRanges object containing a list of NMD-inducible stop codons
}
\description{
This function will check for NMD features for an input transcript
}
\examples{

# Return a GRanges Object with 1 stop codon
testTxforNMD(ptbp2_testData$noNMD, mmus_dna)

# Return a GRanges Object with multiple premature stop codons, 
# 5 of which is >50 from lastEJC
testTxforNMD(ptbp2_testData$NMD, mmus_dna)

}
\author{
Fursham Hamid
}
