% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_NMD.R
\name{testTXforStart}
\alias{testTXforStart}
\title{Test transcripts for annotated start codons}
\usage{
testTXforStart(knownCDS, queryTX)
}
\arguments{
\item{knownCDS}{A GRanges object containing CCDS information of a gene family}

\item{queryTx}{A GRanges object containing exon structure of a query transcript from
the same gene family}
}
\value{
A list of GRanges objects similar to 'indentifyAddedRemovedRegions()' output. 
Fourth object in the list flags the transcript for annotated start codon
?indentifyAddedRemovedRegions for more information
}
\description{
This function will test whether the query transcript contain an annotated 
Start codon from a reference CCDS of the gene
}
\examples{

testTXforStart(ptbp2_testData$noNMD, ptbp2_testData$NMD)


}
\author{
Fursham Hamid
}
